Remove duplicate bbox filter from GetFeatureInfo query
This duplicate bbox in the SQL query is harmless in most situations,
but not with Oracle. This DB seems to have a bug and don't use
the spacial index if there is two SDO_FILTERS in the where clause
that are fed using bound variables (reproduced in a small java
program that is uses only ojdbc).