Correct errors in the junit test code.
The code that tests the fix for GEOT-5239 (connection unwrapping) was
incorrectly located in the JDBCConnectionLifecycleOnlineTest which
was inappropriate given the scope is limited to Oracle support.  That
test code has been relocated to OracleConnectionLifecycleOnlineTest.

Also apply a fix for GEOT-5248 (CSVWriteTest fails). Andrea confirms
that this works on the Windows 10 server build (see jira history).
But this consistently fails on my Windows 7 workstation.  The failure
happens because the remote baseline text file was pulled down without
\r in the line ending but the locally written file to compare with
that baseline does have \r in the line endings.  My change is to strip
\r from both of them before doing the compare.  See GEOT-5248 for more
discussion of this and the relation to git configuration I'm using
which conforms to geotools documentation.