Avoid ClassCastException in raster rescaling.
RenderedOp.getData() is defined to return a Raster. It often returns a
WritableRaster, but in some cases (the ones I saw were for 32 bit float
rasters with an odd width) it returns a plain read-only Raster.

Since we need a WritableRaster, in those cases we call copyData()
instead. copyData() is defined to return a WritableRaster, but
presumably copies data around and therefore might have a memory
or performance cost, so we don't want to do it if getData() works.

GEOT-5463: ClassCastException in imagemosaic-jdbc